Adds a model for the resource schema (#12)
* Adds a model for the resource schema

ResourceTypeSchema introduced to express simple modeling of the extended meta schema properties we have. This type extends the ObjectSchema in the everit-org/json-schema package and converts the unprocessed schema fields (those extensions which are not part of draft-07) into modeled properties. Would prefer to use a consistent Builder pattern but some of the types are internal and not easily extended. What I have here is simpler anyway, and achieves the result.

A Parser shim on top of the model converts types like JSONPointer into String for simpler parsing. Not strictly necessary but might be helpful.

Extended the test schema to allow tests to verify all the extended property modeling.